+++ versions.tf
@@ -5,5 +5,10 @@ terraform {
       source  = "hashicorp/azurerm"
       version = ">= 2.48"
     }
+    # tflint-ignore: terraform_unused_required_providers
+    azurecaf = {
+      source  = "aztfmod/azurecaf"
+      version = "~> 1.1"
+    }
   }
 }